Added CreateVoteModal and integrated it with inlineVote component
